requirement.txt
pyrogram
tgcrypto
requests
beautifulsoup4
motor
pymongo
python-dotenv